requirements.txt
line-bot-sdk
flask
openai
google-genai
gunicorn
